Refactor core, ecdsa, tbtc and tbtc-dapp into their own script
When we restart `geth`, we need to re-deploy all of our contracts.
Unfortunately, running `install.sh` is going to try to refetch master
which may not be what we want (we might be making changes) and also
takes time.

Instead, we refactor the installation of core, ecdsa, tbtc and tbtc-dapp
into their own `install-applications-with-contracts.sh` script which is
invoked from `install.sh`. This allows us to simply run
`install-applications-with-contracts.sh` if `geth` was restarted at some
point.