Refactor installation to use npm links
The installation scripts can work with npm links instead of providing
addresses of dependent external contracts in files like `external.js`.

Artifacts directory is created in installation script so we don't need
to create it in the scripts in local-setup.

This code depends on couple of other PRs in the modules:
  - https://github.com/keep-network/keep-core/pull/2452
  - https://github.com/keep-network/keep-ecdsa/pull/776
  - https://github.com/keep-network/tbtc/pull/790