Additional flags passed to geth run command
Added some flags that we noticed are stabilizing geth 1.10:

rpc.allow-unprotected-txs - this helps for transaction submitted to geth
that are signed according to the old pre-berlin-hardfork rules. EIP-155
changed way of signing transaction. We already aligned client code in
https://github.com/keep-network/keep-common/pull/72 but our web3 and
truffle scripts are still experiencing some problems. As geth 1.10 turns
on the validation by default we want to switch it off until our codebase
works with it

snapshot=false - we turn off snapshot mechanism for geth that consumes a
lot of resources on the machine during the snapshot preparation stage
but doesn't give us much profit

nodiscover and maxpeers 0 - we're running single node private chain and
we don't need the node to connect to any other nodes, we noticed
problems where the node was reporting connection and synchronisation of
chain state locally and in the CI from another node. It was unclear to us
what the heck is this node connecting as it's supposed to be just the only
instance running in isolation.