Move E2E tests workflow from CircleCI to GH Actions
With this commit/PR we are switching from using CircleCI for performing
of E2E tests to Github Actions. The old CircleCI workflow used for E2E
testing had been broken for some time, so apart from just moving the
steps "one to one" from CCI to GHA, failing steps needed to be fixed.
Some of the changes include:
- creation of config file for GH Actions workflow performing E2E tests
- removal of config file for CircleCI job performing E2E tests
- switching to higher version of Node.js for E2E tests (from `14.3.0`
  to `14.17.0`)
- adjusting steps of the install scripts to changes introduced in
  https://github.com/keep-network/local-setup/pull/76