Combine steps installing Go and setting GOPATH env variable
Adding GOPATH to the list ov environment variables available to GITHUB
can be performed in the same step in which Go gets installed.
However, we cannot include in that step the command which appends the
`$GITHUB_PATH` with `$GOPATH/bin` value, as the `$GOPATH` variable would
not be available at this stage yet.