Run e2e testnet tests on Ropsten via GHA, not Kubernetes
We were running E2E tests on testnet (Ropsten) using Kubernetes. But as
the GitHub Actions provide better visibility for the tests results, we
decided to move the testing job there.
The `e2e-testnet.yml` has been created and the old Kubernetes config has
been removed.
The currently proposed solution is not ideal. It requires manual
updating of the `tbtc.js` version in the workflow config every time
new contracts get released. In the future we plan to do that
automatically.