Make sed invocations cross-platform
Syntax of the `sed -i` command is different for MacOS and for GNU
systems. The command in its old shape was failing when run on Ubuntu
(during execution of workflow with E2E tests). Fixed by merging the
problematic command with another sed command in the same file replacing
the text by using temporary file (instead of replacing in the current
file).