Set up Go 1.16 using setup-go action
Upgrading version of Go used in E2E tests local scripts to 1.16.6. With
previously used version building of keep-ecdsa client was failing with
`note: module requires Go 1.16` error.
It turned out that just changing the version of Go in the
`install-go.sh` script was not sufficient, we had to rework the way
we're installing Go.
With previous way of installing Go, the 1.15.13 version was always used,
no matter what version was specified in the `install-go.sh` script. This
likely was caused by the fact that Go 1.14.15, 1.15.13 and 1.16.5 are
the cached tools of the `ubuntu-latest` virtual environment.
We are now switching to use of the `setup-go` action to force using of
1.16 version of Go.
We still need to pass Go-related configs to GITHUB_ENV and GITHUB_PATH
(without this installation of Protobuf fails).