Update manual-setup.adoc after incorporating Coverage Pools
With the introduction of Coverage Pools, the installation process has
changed. We've created `setup-coverage-pools.sh` script which is
executed as part of `install.sh`. We need the documentation for the
manual setup to reflect those changes.
As the installation of the `coverage-pools` requires `Yarn`, `Yarn` has
been added to the list of tools that need to be installed in order to
setup a local machine.