No longer install `sha3` in a separate step
Up until recently, the `package-lock.json` in `keep-ecdsa/solidity` had
the `sha3` package configured in the `1.2.3` version. This version
wasn't working correctly with the Node.js 14.x which we're using for E2E
testing. This meant we needed to upgrade the `sha3` to a higher version.
Previously we've been doing that in the `install-e2e-test.sh` script,
but recently we upgraded the version (to `1.2.6`) directly in the
`keep-ecdsa/solidity/package-lock.json` file, making the `npm install
--save-dev sha3` step in the `install-e2e-test.sh` script obsolete.