Bump up Node.js version used in workflow to 14.17.0
Previously for part of the workflow we were using Node.js in version 12
and for part - version 14.17.0. Version 12 was needed because some of
the installed subprojects were built using v12. Now all subprojects are
built using v14, so we can use one common v14 for the whole workflow
with e2e tests.