Resolve chainId for PrivateKeyWalletSubprovider
PrivateKeyWalletSubprovider expects chainId to be provided as a
constructor argument. If the value is not provided the provider assumes
it deals with mainnet (`chainId = 1`).

After EIP-155 the `chainId` is included when calculating singature, and
transacctions that are send to the Ethereum endpoint with non-matching
`chainId` are rejected with `invalid sender`.

The previous solution worked fine with Ethereum endpoint that had replay
protection disabled, but once the protection is enabled we need to
include the correct `chainId` in the transaction signature.