Increase frequency of running E2E tests on Testnet
After fixing some other issue causing E2E tests to fail we observed
that the test were still failing, this time with `execution reverted:
not at current or previous difficulty` error.
This is an error which we have seen in the past, happening for around
60% of jobs. Now we're seeing it for 100% of jobs, but we haven't run
many jobs so far since fixing that other issue. We want to increase the
frequency of tests execution to see if we will get some passing jobs
apart from those which fail.