Remove `git init` from the build steps
These were necessary back when I was running keep-core's install.sh directly (which assumes it's being run in a git repo).

Over time, I replicated the functionality in install.sh and split it up into the dockerfile and entrypoint for efficiency. We want as much as possible to live at build-time!